+++ app/gui/secrets/decrypt_secrets_gui_screen.go
@@ -10,14 +10,16 @@ import (
 // ScreenDecryptSecrets shows the decrypt secrets screen
 func ScreenDecryptSecrets(window fyne.Window) fyne.CanvasObject {
 	// secrets password
-	passwordEntry := widget.NewPasswordEntry()
+	var secretsFiles = uielements.CreateSecretsFileEntry()
+	var passwordEntry = widget.NewPasswordEntry()
 
-	form := &widget.Form{
+	var form = &widget.Form{
 		Items: []*widget.FormItem{
+			{Text: "Secrets file", Widget: secretsFiles},
 			{Text: "Password", Widget: passwordEntry},
 		},
 		OnSubmit: func() {
-			_ = secretsactions.ActionDecryptSecretsFile(passwordEntry.Text)
+			_ = secretsactions.ActionDecryptSecretsFile(passwordEntry.Text, secretsFiles.Selected)
 			uielements.ShowLogOutput(window)
 		},
 	}